refactoring aplication layers and constants organization

+++ src/user/application.rs
@@ -1,12 +1,13 @@
 use std::error::Error;
-use crate::session::domain::SessionRepository;
 use crate::metadata::domain::{Metadata, MetadataRepository};
+use crate::security;
+use crate::constants::errors;
 use super::domain::{User, UserRepository};
 
-pub fn user_signup<'a>(user_repo: &dyn UserRepository,
-                       meta_repo: &dyn MetadataRepository,
-                       email: &'a str,
-                       password: &'a str) -> Result<(), Box<dyn Error>> {
+pub fn user_signup(user_repo: &dyn UserRepository,
+                    meta_repo: &dyn MetadataRepository,
+                    email: &str,
+                    password: &str) -> Result<(), Box<dyn Error>> {
     
     info!("got signup request from user {} ", email);
     
@@ -15,18 +16,22 @@ pub fn user_signup<'a>(user_repo: &dyn UserRepository,
     Ok(())
 }
 
-pub fn user_delete<'a>(user_repo: &dyn UserRepository,
-                       sess_repo: &dyn SessionRepository,
-                       email: &'a str) -> Result<(), Box<dyn Error>> {
+pub fn user_delete(user_repo: &dyn UserRepository,
+                    email: &str,
+                    pwd: &str,
+                    totp: &str) -> Result<(), Box<dyn Error>> {
     
     info!("got a deletion request from user {} ", email);
-    
+
     let user = user_repo.find(email)?;
-    
-    if let Ok(sess_arc) = sess_repo.find_by_email(email) {
-        if let Ok(sess) = sess_arc.read() {
-            sess_repo.delete(&sess)?;
-        }
+    if !user.match_password(pwd) {
+        return Err(errors::NOT_FOUND.into());
+    }
+
+    // if, and only if, the user has activated the 2fa
+    if let Some(secret) = &user.secret {
+        let data = secret.get_data();
+        security::verify_totp(data, totp)?;
     }
 
     user_repo.delete(&user)?;
